+++ packages/app_package_maker/CHANGELOG.md
@@ -1,3 +1,8 @@
+## 0.3.5
+
+* `MakeConfig` adds `appBinaryName` field.
+* Add `MakeLinuxPackageConfig` class.
+
 ## 0.3.4
 
 * bump `shell_executor` to 0.1.4